refactor progress indicator package

+++ appkit/button/button.h
@@ -3,4 +3,4 @@
 void* Button_New(long id);
 void Button_SetTitle(void* ptr, const char* title);
 void Button_SizeToFit(void* ptr);
-void Button_SetBezelStyle(void* ptr, unsigned int style);
+void Button_SetBezelStyle(void* ptr, unsigned long style);
